Make is_integer a type guard, move to typing
